* commit 'f7d997e592bb0d86ef5ba599ba36e39363340fa6':
  64bit ART object model changes.

+++ libart/src/main/java/java/lang/DexCache.java
@@ -71,7 +71,7 @@ final class DexCache {
     String[] strings;
 
     /** Holds C pointer to dexFile. */
-    private int dexFile;
+    private long dexFile;
 
     // Only created by the VM.
     private DexCache() {}
